change monitor and NC benchmark
